修改 插件列表实现方式抽离成公共 component
