+++ src/views/Admin/Plugins/Add/locale/zh-cn.js
@@ -1,0 +1,3 @@
+export default {
+
+}
