+++ src/views/Admin/Plugins/List/plugin.js
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types'
 import ImmutablePropTypes from 'react-immutable-proptypes'
 
 import { Ul, Li } from 'components/InlineUl'
+import { PluginListItem } from '../components/Plugins'
 
-import classnames from 'classnames'
 import classes from './plugin.scss'
 
-export default class Plugin extends Component {
+export default class PluginController extends Component {
   static propTypes = {
     plugin: ImmutablePropTypes.map.isRequired,
     /**
@@ -52,42 +52,6 @@ export default class Plugin extends Component {
     </Ul>
   }
 
-  renderName () {
-    const { plugin } = this.props
-    const enabled = plugin.get('enabled')
-    return <div className={classes.plugin}>
-      <h5 className={classnames(classes.strong, classes.name)}>
-        {plugin.get('name')}
-      </h5>
-      {this.renderActions(enabled)}
-    </div>
-  }
-
-  renderDesc () {
-    const { plugin } = this.props
-    const tags = plugin.get('tags')
-    return <div className={classes.desc}>
-      <h5 className={classes.strong}>{plugin.get('desc')}</h5>
-      <Ul className={classes.tags}>
-        <Li>版本: {plugin.get('version')}</Li>
-        {tags.map((t) => <Li key={t}>{t}</Li>)}
-        <Li>
-          <a href={plugin.get('link')} target='_blank'>使用帮助</a>
-        </Li>
-      </Ul>
-    </div>
-  }
-
-  renderPlugin () {
-    return <div className={classes.item}>
-      <div className={classes.icon}>
-        <i className='icon icon-jigsaw' />
-      </div>
-      {this.renderName()}
-      {this.renderDesc()}
-    </div>
-  }
-
   renderUpdateInfo () {
     const name = this.props.plugin.get('name')
     return <div className={classes.update}>
@@ -104,12 +68,11 @@ export default class Plugin extends Component {
     const { plugin } = this.props
     const version = plugin.get('version')
     const lastest = plugin.get('lastest')
-    if (lastest && lastest !== version) {
-      return <div className={classes.itemGroup}>
-        {this.renderPlugin()}
-        {this.renderUpdateInfo()}
-      </div>
-    }
-    return this.renderPlugin()
+    const canUpgrade = lastest && lastest !== version
+    const enabled = plugin.get('enabled')
+    return <PluginListItem plugin={plugin}
+      actions={this.renderActions(enabled)}>
+      {canUpgrade && this.renderUpdateInfo()}
+    </PluginListItem>
   }
 }
