+++ src/views/Admin/Plugins/components/Plugins/index.js
@@ -1,0 +1,7 @@
+import PluginList from './plugins'
+import PluginListItem from './plugin'
+
+export {
+  PluginList,
+  PluginListItem,
+}
