+++ src/views/Admin/Plugins/components/Plugins/plugin.scss
@@ -1,0 +1,78 @@
+@import 'variables/colors';
+.plugins{
+  color: $color-black;
+  margin-bottom: 30px;
+}
+.header{
+  display: flex;
+  align-items: center;
+  background-color: #EDEFF1;
+  font-weight: 500;
+  color: $color-black;
+  height: 30px;
+  >div{
+    padding-top: 0px;
+    padding-bottom: 0px;
+  }
+}
+.itemGroup{
+  border-bottom: 1px solid #D7D9DC;
+  .item{
+    border-bottom: none;
+  }
+}
+
+.item{
+  display: flex;
+  align-items: baseline;
+  border-bottom: 1px solid #D7D9DC;
+}
+.update{
+  background-color: #EDEFF1;
+  font-size: 12px;
+  line-height: 30px;
+  padding: 0px 10px;
+  margin-bottom: 10px;
+  a{
+    cursor: pointer;
+  }
+}
+
+.icon{
+  flex: 0 0 30px;
+  padding: 10px 0 10px 10px;
+  text-align: left;
+}
+.plugin{
+  flex: 0 0 auto;
+  width: 30%;
+  padding: 10px 15px;
+}
+.name{
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+.desc{
+  flex: 1;
+  padding: 10px 15px;
+
+}
+.strong{
+  margin-bottom: 5px;
+}
+.actions{
+  font-size: 12px;
+  color: #6668D3;
+  a{
+    cursor: pointer;
+  }
+}
+.grey{
+  color: #8B8C9B;
+}
+
+.tags{
+  font-size: 12px;
+  color: #8B8C9B;
+}
